+++ packages/fsengagement/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [10.18.0](https://github.com/brandingbrand/flagship/compare/v10.17.2...v10.18.0) (2022-01-04)
+
+**Note:** Version bump only for package @brandingbrand/fsengagement
+
+
+
+
+
+## [10.17.2](https://github.com/brandingbrand/flagship/compare/v10.17.1...v10.17.2) (2021-12-07)
+
+
+### Bug Fixes
+
+* **fsengagement:** no-ticket - fix ShopIngredients context ([b22c095](https://github.com/brandingbrand/flagship/commit/b22c0958da40d1b7b818c9ba1344cf2dd8ee7d01))
+
+
+
+
+
+## [10.17.1](https://github.com/brandingbrand/flagship/compare/v10.17.0...v10.17.1) (2021-11-18)
+
+
+### Bug Fixes
+
+* **fsengagement:** context fix for class components in v10 ([109c7b9](https://github.com/brandingbrand/flagship/commit/109c7b9dde0caaa70e1756e2a312bc3ea12a46c2))
+
+
+
+
+
+# [10.17.0](https://github.com/brandingbrand/flagship/compare/v10.16.0...v10.17.0) (2021-11-09)
+
+**Note:** Version bump only for package @brandingbrand/fsengagement
+
+
+
+
+
 # [10.16.0](https://github.com/brandingbrand/flagship/compare/v10.15.0...v10.16.0) (2021-11-05)
 
 **Note:** Version bump only for package @brandingbrand/fsengagement